added token refresh and blacklist

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from models.item import ItemModel
 
 
@@ -22,6 +22,7 @@ class Item(Resource):
             return item.json()
         return {'message': 'Item Not Found'}, 404
 
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -37,12 +38,17 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required(fresh=True)
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
         return {'message': 'Item Deleted'}
 
+    @jwt_required(fresh=True)
     def put(self, name):
         data = Item.parser.parse_args()
         item = ItemModel.find_by_name(name)
@@ -59,5 +65,12 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {'items': [item.name for item in ItemModel.find_all()],
+                'message': 'More data available if you log in.'
+                }
